chore: type cleanup
