+++ src/models/Game.ts
@@ -1,3 +1,5 @@
+import { Category } from "./Quiz";
+
 export type Stage = 'INIT' | 'GAME' | 'END';
 export type QuestionNumberOptions = 10 | 15 | 20;
 export type QuestionDifficultyOptions = 'easy' | 'medium' | 'hard' | 'any';
@@ -9,4 +11,14 @@ export interface GameSettings {
   difficulty: QuestionDifficultyOptions
   category: QuestionCategoryOptions
   type: QuestionTypeOptions
+}
+
+export interface GameState {
+  stage: Stage;
+  difficulty: QuestionDifficultyOptions;
+  questionNumber: QuestionNumberOptions;
+  questionType: QuestionTypeOptions;
+  questionCategory: QuestionCategoryOptions;
+  categories: Category[];
+  categoriesLoading: boolean;
 }
