+++ src/views/ScorePage.tsx
@@ -2,12 +2,12 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 import Button from '../components/Button';
 import { Answer } from '../models/Quiz';
-import { RootState, useAppSelector } from '../store';
+import { useAppSelector } from '../store';
 import { setStage } from '../store/features/game';
 
 const ScorePage: React.FC = () => {
   const dispatch = useDispatch();
-  const { answers, score, questions } = useAppSelector((state: RootState) => state.quiz);
+  const { answers, score, questions } = useAppSelector((state) => state.quiz);
 
   const handleRestart = (): void => {
     dispatch(setStage('INIT'));
